Switched heuristic return value to float

+++ include/heuristics.h
@@ -5,12 +5,12 @@
 
 namespace Heuristics {
     // Utility function
-    int evaluateEndState(uint8_t** board, uint8_t playerNumber);
+    float evaluateEndState(uint8_t** board, uint8_t playerNumber);
 
     uint8_t playerRanking(uint8_t **board, uint8_t playerNumber);
 
     // Heuristic
-    int weightedHeuristic(uint8_t** board, uint8_t playerNumber);
+    float weightedHeuristic(uint8_t** board, uint8_t playerNumber);
 
     int getScore(uint8_t** board, uint8_t playerNumber);
     int getScoreEnemyMoves(uint8_t** board, uint8_t ourPlayerNumber);
